URP 2D: Fix backbuffer in overlay cameras
What is happening is that prior to RTHandles, getTemporaryRT could be identified by Shader.PropertyToId(name).
However, using RTHandles, the RTID is now more complex.

This means that an overlay camera trying to get such a handle using PropertyToId will not map to the resource.
A similar issue anticipated and resolved in UniversalRenderer.