[VFX] Instancing: First iteration (v2)
VFX Instancing is a feature for VFX Graph that allows batching VisualEffects in memory to perform some operations on them at the same time.

This is the first iteration of the feature, which we consider incomplete at the moment. One or more PRs related to this will need to land before 2022.2.

* How it works
In general, the process is automatic and VisualEffects will be batched together if the feature is enabled. We included some options to control this behavior:
- The feature can be globally disabled in Editor in developer mode, using an option under "Preferences> Visual Effects> Instancing enabled".
- Each VFX graph asset has an option in the inspector to control how the VFX of that asset type are handled. That option allows to disable instancing for that particular asset, to automatically decide the batch size(default), or to manually specify how many VFX can be per batch.
- Once the VFX is created, the VFX Renderer allows to disable only the rendering part of the feature

Some VisualEffect assets will have instancing forced to disabled if:
- A particle system has an output node that is transparent.
- Uses sorting.
- Use any ShaderGraph material.
- Has particle strips
- Uses output events
- Uses GPU events
- Has Mesh outputs
- Exposes variables that are Gradient, Curve, Texture or GraphicsBuffer

* What needs to be tested:
The feature involved an important refactor, so any VFX can be affected, even if instancing is disabled. We have a good test coverage but it would need additional general testing in case we are missing something.
Being a first iteration, we are mostly interested in stability. For the next iteration we will test more performance and usability.

* JIRA:
https://jira.unity3d.com/browse/VFXG-182

* Design document:
https://docs.google.com/document/d/1WmsOZp1zd3Gl1u7HatiEh-AgCyTp90a_O1RRJPAD9v4

* Short technical high-level presentation on changes to the structure:
Slides: https://docs.google.com/presentation/d/1Ddmfk0Gbn0gSt-DV1PZcPw_qhN_MkH3sIekGuiwBEHk/edit?usp=sharing