[URP] Case 1418462: Improved fallback to single shadow cascade on GLES2
Fix for ISS case https://fogbugz.unity3d.com/f/cases/1418462/

We strip shadow cascade keywords from the lighting shaders on GLES2. This was not communicated to users and can cause confusion.

This PR adds an info box under the shadow cascades settings, when targeting GLES2, notifying the users about the fallback to 1 cascade.
It also makes sure that we fallback correctly to 1 cascade at runtime too, so that the shadows max distance is correctly taken into account.