shader replacement for SRP's (v3)
https://jira.unity3d.com/browse/RND-175
https://jira.unity3d.com/browse/GFXRPF-27
https://docs.google.com/document/d/1x8IqIut5iANYzZSbUpfZWeFn-HymlpbGsYEkj2vQo5Y/edit?usp=sharing
In Built-in there is Camera.RenderWithShader(). For SRP's this functionality does not exist. This PR adds a replacement shader parameter to the DrawRenderers command.

There is already an overrideMaterial available but there is a key distinction between overrideMaterial and replacementShader.
overrideMaterial: override the material and all material properties.
replacementShader: override the shader and keep all current properties and try to fit them to the replacement shader. This way users can access textures and properties originally on the renderer in their replacementShader.

This is mainly useful for debug views or other effects that need to access things such as alpha cut-out textures.

This PR also adds UI so this feature is available in the DrawObjects and DrawRenderers customPass and RenderFeature in HDRP and URP.

https://drive.google.com/file/d/1kASDA2S95MijYMpihLT2sWQpzEj8wbiq/view?usp=sharing


This feature is not yet supported by the SRPBatcher and BatchRendererGroups. This will be scheduled in the next release and has been added to our backlog for tracking purposes. https://jira.unity3d.com/browse/GFXRPF-40