[URP] Automatic Alpha-To-Coverage Support
This change attempts to automatically enable alpha-to-coverage
functionality whenever alpha clipping is in use on opaque materials
in URP. Since this functionality relies on MSAA, it is only active for
forward passes. With this change, alpha clipped geometry such as
foliage should see improved visual quality when MSAA is enabled.

No MSAA:
![image](https://user-images.githubusercontent.com/89797527/142033665-071df79a-44ca-4145-a299-b5452a8e10a2.png      )

4x MSAA:
![image](https://user-images.githubusercontent.com/89797527/142033684-c1f24950-3fb7-4c4e-9ed5-95379541a16d.png      )

No MSAA (5x Zoom):
![image](https://user-images.githubusercontent.com/89797527/149583024-0e7e094e-a658-4dd2-b100-67d0bd9fd5b4.png      )

4x MSAA (5x Zoom):
![image](https://user-images.githubusercontent.com/89797527/149583035-878397d1-2e18-4f64-a960-79f489620784.png      )