Ref optimizations for URP light loops.
Reduce the amount of memcpys in URP VisibleLight loops code resulting from var tmp = NativeArray[i]; style code.
This PR adds an "unsafe" extension methods for NativeArrays in URP. They allow to return the array index as a reference instead of a copy.
The references in loops are only used for readonly purposes, therefore the accesses are actually safe.