Reworking lighting features
Added more remaps using the BRDFMAP. Added more anime specific features, you'll have to define ANIME in your shader to use them. Swapped out the shadow biasing with what we used in Zero Lab Renderer as the URP one was nearly useless due to the heavy artifacting. Swapped out the 5x5 shadow filter with the 7x7 version for better soft shadows. This is a per light option in URP 14 which also gets rid of the shader variants. We should update URP to better account for this.