Volumetric options
Moving where extinction is added to the system so we can more easily add and modify to it. It's ready for a precomputed density pass to be added to the clipmap. Added a height density option in the volume and added a base cubemap contribution based on the same mipfog to avoid over darkening. This can act as the base fog too so you may not need that anymore. depending on the use case.