[XR] [URP] [2022.2] (AR Foundation) Fix forward opaque color clear flag being applied when before opaque passes exist
URP 14 currently causes problems with RendererFeatures that render to the render target prior to rendering opaques. AR Foundation is a Unity product that is negatively affected by this issue. This is seemingly due to the configuration of the `renderOpaqueForwardPass.ConfigureClear` on base cameras without accounting for having before opaque passes. This PR enforces the clear flag None when the active render pass queue finds a before opaques pass as the render target has valid rendering output.