[2022.3] [URP] Ensure HDR output can be turned off by script in the editor
HDR Output is automatically turned on (by the URP) in the editor if available, so the game view can be the closest to the final result once built and run on an HDR display. However it would turn HDR on even if the users explicitly disabled it during playmode.
This PR ensures the URP takes into account calls to `HDROutputSettings.main.RequestHDRModeChange(false)` in the editor, by enabling HDR output only once when it detects it as available for the first time.