[Port] [2022.3] URP Decals normal blending issues
In URP, the combination of the Deferred rendering path, Accurate G-Buffer Normals, and the Screen Space decal technique, means that normal blending is not supported. This is because the Screen Space decal technique in combination with Deferred blends directly into the normals buffer, but the octahedral format used for Accurate G-Buffer Normals does not support blending due to discontinuities.

This PR also changes the Automatic decal technique option to take this into account. Previously, it would pick D-Buffer on non-mobile / non-OpenGL, and Screen Space otherwise. It now takes the Accurate G-Buffer Normals option into account, preferring D-Buffer if the option is enabled since Screen Space would yield normal blending issues.

Also worth noting is that previously, the Automatic option would never pick the G-Buffer technique for Deferred. The UI only exposes Automatic, D-Buffer, and Screen Space, with the last one mapping to G-Buffer if deferred. So the user can't force the actual Screen Space technique for Deferred, but the Automatic option will use it anyway. I changed that so that the Automatic option now uses the G-Buffer approach if deferred in the cases where it would previously use the actual Screen Space technique.