Fixing vertex deformations when using scene selection and picking in combination with the BRG
When using BRGs(such as with entities) picking and selection outline would not correctly reflect vertex deformations. This is because the DOTS_INSTANCING wasn't defined in the relevant passes. This PR adds DOTSPregraph to passes so that the DOTS_INSTANCING marcro is defined when needed.
Ticket ID: https://jira.unity3d.com/browse/UUM-51502