[URP] Fixed FXAA with point upscaling when HDR output is active
The FXAA algorithm relies on values being in the range [0, 1], but the HDR output processing require to pass values in nits in the maximum theoretical range [0, 10k]. So when working with an HDR-output image, we need to adjust the dynamic range before and after applying the FXAA. We ensure post-processing shaders that work with dynamic range know we are using HDR output through the `HDR_INPUT` keyword.

But when using the old FXAA implementation (2022.3 and before), in combination with nearest-neighbour or bilinear upscaling, one of the shader (`ScalingSetup.shader`) during Final Post Processing was not defining the `HDR_INPUT` keyword. This caused the dynamic range of the image to be clamped to [0, 1] nits by the FXAA, and appear very dark on an HDR monitor.

This PR fixes the issue by properly enabling the necessary `HDR_INPUT` keyword. The previous `_GAMMA_20_AND_HDR_INPUT` keyword has been removed and replaced by `_FXAA_AND_GAMMA_20`. It is meant to reduce shader variant count since it makes more sense for FXAA/FSR keywords to be combined than HDR output keywords:
Before it had 2 set of exclusive keywords: "no FXAA"/"FXAA" and "no encoding"/"Gamma 2.0"/"Gamma 2.0 + HDR input". This produced 6 variants of the pixel shader, including one that wasn't used ("no FXAA" + "no encoding", since the ScalingSetup pass is only used if either FXAA or FSR or both is enabled.) When HDR outpu is disallowed, we fall down to 4 variants.
Now there's still 2 set of exclusive keywords: "FXAA"/"Gamma 2.0"/"FXAA + Gamma 2.0" (note there's no default shader variant that doesn't do either) and "SDR input"/"HDR input". This still produces 6 variants but when HDR output is disallowed we fall down to 3 variants.

The `HDR_INPUT` variant is still stripped when HDR output is not allowed, by `HDROutputUtils.IsShaderVariantValid`.