URP: Improved renderViewportScale for viewspace intermediate textures.
This change makes sure that tempRT are scaled correctly with Quest renderViewportScale feature without having to reallocate them.

Currently, changing XRSettings.renderViewportScale will trigger a reallocation of all tempRT since we modify the cameraTargetDescriptor width/height, which is used by render passes to configure their tempRT.

The fix avoids changing the value of cameraTargetDescriptor by using the full RT size as the width and height and enabling scaling of all render targets (using the ScalableBufferManager) that use the cameraTargetDescriptor.