[UUM-60556][PIA-1211][PIA-1189] 2022.3 Fix URP sRGB conversion
Partial backport of https://github.cds.internal.unity3d.com/unity/unity/pull/25124

The backport fixes the sRGB conversion when PostProcessing is turned off (either globally or on the camera). In this case the `FinalBlit` path was used, which had no sRGB conversion in the blitting shaders.
It also fixes debug views, which also not worked with PostProcessing turned off.

The backport is needed for one of our customers (Stellantis) who reported that the screen is to dark when it is running on a device with a gamma space backbuffer. This was because the rendering path lacked a color conversion from the sRGB temp buffer to the backbuffer.