[Port] [2022.3] [URP] Fix in use renderer destruction from other UniversalRPAsset inspector
Some fields use shared data through different renderer instances.

Was not able to produce it on my computer at all, but was able to produce it on Kirill ones and we pair debugged it on 6000. The root of issue is in the end that the Renderer is shared between the RPAsset in use and the RPAsset edited through the inspector. To assess the compatibility of the renderer, it need to be constructed again with the current architecture. And then when the inspector is modified we destroy the constructed copy. It could have work but they share some data that got destroyed and caused the null reference if the rendering occures between shared data is recreated.

Note that the architecture is as is since the implementation of Renderer as an asset. So changing anything here is really touchy. Aggregated test seems to be green but this _require extended QA tests_.