Docs/graphics/17 april quality speint oleks 2022 backport
Backport of the following PR (with some 2022-specific changes):
https://github.cds.internal.unity3d.com/unity/unity/pull/48736

DOCG-5296. "a combination of 2D and 3D Lights and 2D and 3D Renderers in a single Scene"

DOCG-4772. Unlit shaders edit.

DOCG-5083 cameraOutput is not a property of UniversalAdditionalCamera…

DOCG-4935 There should be {} at the end of the property after "white"