[Port][2022.3][URP] SingleCameraRequest : Add missing RenderPipelineManager Begin/End Camera and Context callbacks
SingleCameraRequest on URP are rendering a single camera, it's purpose is rendering a camera without the full list of stack cameras. This avoid some overhead.

As we are rendering 1 single camera, we are missing:

```
RenderPipelineManager.beginContextRendering
RenderPipelineManager.endContextRendering
RenderPipelineManager.beginCameraRendering
RenderPipelineManager.endCameraRendering
```
Jira Issue: https://jira.unity3d.com/browse/UUM-69268

I've added a new docs page in URP with an exaple of how to use SubmitRenderRequest and how to subscribe to RenderPipelineManager callbacks in each case.