[Backport][2022.3][SRPF] Filter Objects in Motion Vector pass by object layer mask
In the Universal Renderer Data, you can select Layer Masks to filter out from the renderer.
For example, you can click URP Renderer > Opaque Layer Mask > Deselect "Water" layer.
If you add a GameObject to that "Water" layer so that the Renderer filters it during rendering, the Motion Vectors for that object are still included in the URP Motion Vector pass.

To avoid that we add Opaque Layer Mask to the MotionVectorRenderPass.